requirements.txt
aiosqlite>=0.20.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
Appium-Python-Client==3.2.0
attrs==25.3.0
beautifulsoup4==4.14.2
Brotli==1.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
dataclasses-json==0.6.7
ddgs==9.6.0
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.118.0
frozenlist==1.7.0
git-filter-repo==2.47.0
google-generativeai==0.8.5
greenlet==3.2.4
groq==0.32.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-anthropic==0.3.21
langchain-community==0.3.30
langchain-core==0.3.76
langchain-groq==0.3.8
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langdetect==1.0.9
langgraph==0.6.8
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.31
lxml==6.0.2
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.3
openai==1.109.1
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pillow==11.3.0
pip==25.2
playwright==1.55.0
playwright-stealth==2.0.0
primp==0.15.0
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
pydantic-settings==2.11.0
pyee==13.0.0
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
selenium==4.22.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.48.0
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.25.0